Extract TaskerTransporter function into a class
